add new file for beam search
